Pass to make `CellShare` faster. 

It is substantially faster than it was previously, but there may still be room for improvement. 

I mainly changed `sharing_components.rs` to make get faster sharing. I can explain synchronously the changes that I made. 

Edit: here are some times. Unfortunately gemm6 adn gemm8 still take really long on debug. 

Weirdly, gemm3 takes longer than gemm4 and gemm6 takes longer than gemm 8. 

Baseline  vs. Current Times for CellShare Pass to Run 
Debug:
gemm2: .151 s | .107 s 
gemm3: 16.523 s  | 1.89 s 
gemm4: 5.222 s  | .96 s 
gemm6: > 1 minute  | 60 s 
gemm8: > 1 miniute | 28 s 
Release:
gemm2: .013 s | .011 s 
gemm3: .755 s | .138 s 
gemm4: .261 s  | .083 s 
gemm6: > 1 minute | 3.022 s 
gemm8: > 1 minute  | 1.48 s