@rachitnigam 2022-07-21
Can we get some benchmarking numbers on the difference? For example, how long does the debug and release builds take on the gemm kernels in the benchmarks/ folder