@rachitnigam 2022-07-21
Thanks, this is really cool! We should keep working on it but I think these numbers are good enough to merge these changes. It would be interesting to figure out why `gemm3` and `gemm6` are so much worse