@rachitnigam 2022-07-21
There is a button on top right to trigger a github action when that happens!