@calebmkim 2022-07-21
I think I fixed the bug. This does make the timing (slightly) worse, taking 10 - 25% more time (so at least the same order of magnitude) for the ones I tested. For example. 
debug gemm3 goes from 1.89 -> 2.1
debug gemm4 goes from .96 -> 1.07
release gemm6 goes from 3.022 -> 3.9
release gemm8 goes from 1.48 -> 1.6