@calebmkim 2022-07-22
So here's why `gemm3` takes longer than `gemm4` ( and `gemm6` takes longer than `gemm8`). 

Basically, for some reason the structure some of the big par blocks in `gemm3` is like this: 
```
par {
                if and1.out with cond2 {
                  seq {
                    let2;
                    upd1;
                  }
                }
                if and3.out with cond3 {
                  seq {
                    let3;
                    upd2;
                  }
                }
                [more if stmts like this]  
```
and the structure for some of the big gemm4 par blocks is like this: 
```
par {
                if and74.out with cond31 {
                  upd80;
                }
                if and77.out with cond32 {
                  upd81;
                }
              [and more if statements like this] 
```
So basically, there are more groups running in parallel in `gemm3` than `gemm4`, which requires more work to determine the conflicts. The same is true for `gemm6` and `gemm8`. 
I wonder why the code generated in `gemm3` has this different structure compared to `gemm4`.