@rachitnigam 2022-07-22
These changes look good to me! Left one comment about code style. Another random thought: do we even need the generic stuff in sharing_components anymore? cell_share is the only thing that implements sharing related reasoning anymore so maybe there is benefits to be had by specializing things for it.

I'm also wondering what it would take to scale up the pass significantly more. The key bottleneck seems to be big `par` blocks that have all-to-all conflicts. Maybe worth looking into how we can rearchitect `ScheduleConflicts` or something to make such conflicts much faster to analyze