@calebmkim 2022-07-22
Yeah I agree it seems like the `cell_share` and `sharing_components` could just be one file. And when we do so, we might be able to see some ways to increase performance. 

I'm also thinking about how we can restructure. I have some ideas in my head but I'm not sure if they would actually make the pass faster.